refactor: 최상단에서 route 관리하기 위한 수정
