--- src/components/community/Board.module.css
@@ -1,3 +1,0 @@
-.board {
-  width: 100%;
-}
