--- src/components/community/boardTitle/BoardTitle.js
@@ -1,7 +1,0 @@
-import styles from "./BoardTitle.module.css";
-
-function Title({ text }) {
-  return <div className={styles.title}>{text}</div>;
-}
-
-export default Title;
