--- src/components/community/boardTitle/BoardTitle.module.css
@@ -1,11 +1,0 @@
-@import "../../../styles/variables.module.css";
-
-.title {
-  font-size: 25px;
-  font-weight: medium;
-  width: 955px;
-  height: 35px;
-  border-bottom: 1px solid var(--border-color1);
-  padding-bottom: 20px;
-  margin-top: 55px;
-}
